Chip away at measurement panel decoupling
