+++ platform/viewer/src/appExtensions/MeasurementsPanel/index.js
@@ -1,7 +1,18 @@
 import ConnectedMeasurementTable from './ConnectedMeasurementTable.js';
+import init from './init.js';
 
+/**
+ * MUST BE REGISTERED _AFTER_ Cornerstone Extension
+ *
+ */
 export default {
+  /**
+   * Only required property. Should be a unique value across all extensions.
+   */
   id: 'measurements-table',
+  preRegistration(configuration = {}) {
+    init(configuration);
+  },
   getPanelModule() {
     return {
       menuOptions: [
